+++ plone/dexterity/filerepresentation.py
@@ -30,6 +30,10 @@ from zope.schema import getFieldsInOrder
 from zope.size.interfaces import ISized
 import tempfile
 
+from AccessControl.class_init import InitializeClass
+from AccessControl import ClassSecurityInfo
+from Products.CMFCore import permissions
+
 
 class DAVResourceMixin(object):
     """Mixin class for WebDAV resource support.
@@ -38,6 +42,9 @@ class DAVResourceMixin(object):
     delegate to more granular adapters.
     """
 
+    security = ClassSecurityInfo()
+
+    @security.protected(permissions.View)
     def get_size(self):
         """Get the size of the content item in bytes. Used both in folder
         listings and in DAV PROPFIND requests.
@@ -54,6 +61,7 @@ class DAVResourceMixin(object):
             return size
         return 0
 
+    @security.protected(permissions.View)
     def content_type(self):
         """Return the content type (MIME type) of the tiem
         """
@@ -62,16 +70,19 @@ class DAVResourceMixin(object):
             return None
         return readFile.mimeType
 
+    @security.protected(permissions.View)
     def Format(self):
         """Return the content type (MIME type) of the item
         """
         return self.content_type()
 
+    @security.protected(permissions.View)
     def manage_DAVget(self):
         """Get the body of the content item in a WebDAV response.
         """
         return self.manage_FTPget()
 
+    @security.protected(permissions.View)
     def manage_FTPget(self, REQUEST=None, RESPONSE=None):
         """Return the body of the content item in an FTP or WebDAV response.
 
@@ -110,6 +121,7 @@ class DAVResourceMixin(object):
 
         return reader.read()
 
+    @security.protected(permissions.ModifyPortalContent)
     def PUT(self, REQUEST=None, RESPONSE=None):
         """WebDAV method to replace self with a new resource. This is also
         used when initialising an object just created from a NullResource.
@@ -166,6 +178,9 @@ class DAVCollectionMixin(DAVResourceMixin):
     delegate to more granular adapters.
     """
 
+    security = ClassSecurityInfo()
+
+    @security.protected(permissions.AddPortalContent)
     def MKCOL_handler(self, id, REQUEST=None, RESPONSE=None):
         """Handle "make collection" by delegating to an IDirectoryFactory
         adapter.
@@ -177,6 +192,7 @@ class DAVCollectionMixin(DAVResourceMixin):
             )
         factory(id)
 
+    @security.protected(permissions.AddPortalContent)
     def PUT_factory(self, name, contentType, body):
         """Handle constructing a new object upon a PUT request by delegating
         to an IFileFactory adapter
@@ -186,6 +202,7 @@ class DAVCollectionMixin(DAVResourceMixin):
             return None
         return factory(name, contentType, body)
 
+    @security.protected(permissions.ListFolderContents)
     def listDAVObjects(self):
         """Return objects for WebDAV folder listings.
 
@@ -218,6 +235,7 @@ class FolderDataResource(Implicit, Resource):
     """
 
     __dav_collection__ = 0
+    security = ClassSecurityInfo()
 
     def __init__(self, name, parent):
         self.__dict__.update({'__parent__': parent, '__name__': name})
@@ -264,22 +282,26 @@ class FolderDataResource(Implicit, Resource):
         """
         return self.__name__
 
+    @security.protected(permissions.View)
     def HEAD(self, REQUEST, RESPONSE):
         """HEAD request: use the Resource algorithm on the data of the
         parent.
         """
         return Resource.HEAD(self.__parent__, REQUEST, RESPONSE)
 
+    @security.protected(permissions.ListFolderContents)
     def OPTIONS(self, REQUEST, RESPONSE):
         """OPTIONS request: delegate to parent
         """
         return self.__parent__.OPTIONS(REQUEST, RESPONSE)
 
+    @security.protected(permissions.View)
     def TRACE(self, REQUEST, RESPONSE):
         """TRACE request: delegate to parent
         """
         return self.__parent__.TRACE(REQUEST, RESPONSE)
 
+    @security.protected(permissions.View)
     def PROPFIND(self, REQUEST, RESPONSE):
         """PROPFIND request: use Resource algorithm on self, so that we do
         not appear as a folder.
@@ -288,6 +310,7 @@ class FolderDataResource(Implicit, Resource):
         """
         return super(FolderDataResource, self).PROPFIND(REQUEST, RESPONSE)
 
+    @security.protected(permissions.ModifyPortalContent)
     def PROPPATCH(self, REQUEST, RESPONSE):
         """PROPPATCH request: Use Resource algorithm on self, so that we do
         not appear as a folder.
@@ -296,16 +319,19 @@ class FolderDataResource(Implicit, Resource):
         """
         return super(FolderDataResource, self).PROPPATCH(REQUEST, RESPONSE)
 
+    @security.protected(permissions.ModifyPortalContent)
     def LOCK(self, REQUEST, RESPONSE):
         """LOCK request: delegate to parent
         """
         return self.__parent__.LOCK(REQUEST, RESPONSE)
 
+    @security.protected(permissions.ModifyPortalContent)
     def UNLOCK(self, REQUEST, RESPONSE):
         """UNLOCK request: delegate to parent
         """
         return self.__parent__.UNLOCK(REQUEST, RESPONSE)
 
+    @security.protected(permissions.ModifyPortalContent)
     def PUT(self, REQUEST, RESPONSE):
         """PUT request: delegate to parent
         """
@@ -319,6 +345,7 @@ class FolderDataResource(Implicit, Resource):
             'folder level instead'
         )
 
+    @security.protected(permissions.DeleteObjects)
     def DELETE(self, REQUEST, RESPONSE):
         """DELETE request: not allowed
         """
@@ -326,6 +353,7 @@ class FolderDataResource(Implicit, Resource):
             'Cannot delete folder data: delete folder instead'
         )
 
+    @security.protected(permissions.AddPortalContent)
     def COPY(self, REQUEST, RESPONSE):
         """COPY request: not allowed
         """
@@ -333,6 +361,7 @@ class FolderDataResource(Implicit, Resource):
             'Cannot copy folder data: copy the folder instead'
         )
 
+    @security.protected(permissions.AddPortalContent)
     def MOVE(self, REQUEST, RESPONSE):
         """MOVE request: not allowed
         """
@@ -340,16 +369,19 @@ class FolderDataResource(Implicit, Resource):
             'Cannot move folder data: move the folder instead'
         )
 
+    @security.protected(permissions.View)
     def manage_DAVget(self):
         """DAV content access: delete to manage_FTPget()
         """
         return self.__parent__.manage_DAVget()
 
+    @security.protected(permissions.View)
     def manage_FTPget(self):
         """FTP access: delegate to parent
         """
         return self.__parent__.manage_FTPget()
 
+    @security.protected(permissions.ListFolderContents)
     def listDAVObjects(self):
         """DAV object listing: return nothing
         """
@@ -798,3 +830,8 @@ class DefaultWriteFile(object):
 
     def flush(self):
         pass
+
+
+InitializeClass(DAVResourceMixin)
+InitializeClass(DAVCollectionMixin)
+InitializeClass(FolderDataResource)